backend/requirements.txt
requests
boto3
pandas
pathlib
langchain
google-generativeai
pydantic
snowflake-connector-python
fastapi
uvicorn
numpy
plotly
python-dotenv
pinecone[grpc]
sentence_transformers
google-search-results
matplotlib
seaborn
dotenv
praw
regex
nltk
serpapi
langgraph
langchain
langchain-google-genai
tavily-python
beautifulsoup4
aiohttp
langgraph
langchain
langchain-google-genai
webdriver-manager
PyMuPDF
chardet
fake-useragent
cloudscraper
playwright

